chore: refactor components and cleanup codebase
This commit includes several maintenance tasks:

- Replaced NewDueDatePicker with the more generic DatePicker component

  in AddTodoForm and EditTodoForm.

- Removed the unused DateRangePicker component.

- Cleaned up unnecessary comments in the DatePicker component.

- Updated unit tests for TodoItem and accessibility tests

  to reflect component changes and refactoring.

- Moved DatePicker to src/components/ui/ and PriorityPicker

  to src/components/ for better file organization.
