--- src/components/date-picker/date-range-picker.tsx
@@ -1,56 +1,0 @@
-import * as React from 'react';
-import { type DateRange } from 'react-day-picker';
-import { addDays, format } from 'date-fns';
-import { CalendarIcon } from 'lucide-react';
-
-import { Button } from '@/components/ui/button';
-import { Calendar } from '@/components/ui/calendar';
-import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
-import { cn } from '@/lib/utils';
-
-export default function DateRangePicker({ className }: React.HTMLAttributes<HTMLDivElement>) {
-	const [date, setDate] = React.useState<DateRange | undefined>({
-		from: addDays(new Date(), -20),
-		to: new Date(),
-	});
-
-	return (
-		<div className={cn('grid gap-2', className)}>
-			<Popover>
-				<PopoverTrigger asChild>
-					<Button
-						id="date"
-						variant="outline"
-						className={cn(
-							'w-[300px] justify-start text-left font-normal',
-							!date && 'text-muted-foreground'
-						)}
-					>
-						<CalendarIcon className="mr-2 h-4 w-4" />
-						{date?.from ? (
-							date.to ? (
-								<>
-									{format(date.from, 'LLL dd, y')} - {format(date.to, 'LLL dd, y')}
-								</>
-							) : (
-								format(date.from, 'LLL dd, y')
-							)
-						) : (
-							<span>Pick a date</span>
-						)}
-					</Button>
-				</PopoverTrigger>
-				<PopoverContent className="w-auto p-0" align="start">
-					<Calendar
-						autoFocus
-						mode="range"
-						defaultMonth={date?.from}
-						selected={date}
-						onSelect={setDate}
-						numberOfMonths={2}
-					/>
-				</PopoverContent>
-			</Popover>
-		</div>
-	);
-}
